Started this class and ran into a problem
GeoPackage does not use the OGC document template so there are no requirements in the correct template format. And there are no URIs associated with each requirement. As such, dependencies cannot be easily specified in a Requirements Class table. I have a question into Scott as to what to do.